Fix build error

+++ test/conformance/enqueue/urEnqueueUSMFill2D.cpp
@@ -81,8 +81,8 @@ struct urEnqueueUSMFill2DTestWithParam
                                  pitch, width, height, 0, nullptr, nullptr));
 
         size_t pattern_index = 0;
-        for (int w = 0; w < width; ++w) {
-            for (int h = 0; h < height; ++h) {
+        for (size_t w = 0; w < width; ++w) {
+            for (size_t h = 0; h < height; ++h) {
                 uint8_t *host_ptr = host_mem.data();
                 size_t index = (pitch * h) + w;
                 ASSERT_TRUE((*(host_ptr + index) == pattern[pattern_index]));
